+++ html/8_tab_Diagram.php
@@ -58,7 +58,7 @@ include "8_funktion_Diagram.php";
 # Prüfen ob SQLite Voraussetzungen vorhanden sind
 $SQLite_file = $PythonDIR."/PV_Daten.sqlite";
 if (!file_exists($SQLite_file)) {
-    echo "\nSQLitedatei $filename existiert nicht, keine Grafik verfügbar!";
+    echo "\nSQLitedatei $SQLite_file existiert nicht, keine Grafik verfügbar!";
     echo "</body></html>";
     exit();
 }